jelsaw: make the syntax of <section>/<key> unambiguous
For the view command "alpha" could be either a section or key to print.
Now you'd have to specify "alpha/" if you mean the section.  For the copy
command this is interpreted as "alpha/password" now.

In interactive mode the argument can be left out for these commands, which
means "/" (= "default/").